refactor:コンポーネント名の大文字小文字を統一
