+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
-import Footer from "../components/organisms/footer";
+import Footer from "../components/organisms/Footer";
 import ClientSideComponent from '../components/organisms/ClientSideComponent';
 
 const inter = Inter({ subsets: ["latin"] });
