+++ app/main/page.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import React from 'react';
-import dynamic from 'next/dynamic';
 
 // import DogRandomAnimation from '../components/dogRandomAnimation';
 import DogAnimation from '../../components/animations/dog/DogAnimation';
